fix: login modal close
