+++ apps/next/src/pages/_app.tsx
@@ -24,7 +24,7 @@ import { ClaimScreen } from "app/screens/claim";
 import { ClaimLimitExplanationScreen } from "app/screens/claim-limit-explanation";
 import { CollectorsScreen } from "app/screens/collectors";
 import { CommentsScreen } from "app/screens/comments";
-import { CompleteProfileScreen } from "app/screens/complete-profile";
+import { CompleteProfilePromptScreen } from "app/screens/complete-profile-prompt";
 import { DetailsScreen } from "app/screens/details";
 import { DropScreen } from "app/screens/drop";
 import { DropEventScreen } from "app/screens/drop-event";
@@ -188,7 +188,7 @@ function App({ Component, pageProps, router }: AppProps) {
 
         {/* Settings that renders on top of other modals */}
         <EditProfileScreen />
-        <CompleteProfileScreen />
+        <CompleteProfilePromptScreen />
         <AddEmailScreen />
         <VerifyPhoneNumberScreen />
