+++ packages/app/components/complete-profile-modal-content/complete-profile-modal-content.tsx
@@ -1,5 +1,4 @@
 import * as React from "react";
-import { Platform } from "react-native";
 
 import { Button } from "@showtime-xyz/universal.button";
 import { useRouter } from "@showtime-xyz/universal.router";
@@ -38,24 +37,7 @@ export const CompleteProfileModalContent = ({
       <Button
         tw="mt-12"
         size="regular"
-        onPress={() =>
-          router.push(
-            Platform.select({
-              native: "/profile/complete",
-              web: {
-                pathname: router.pathname,
-                query: {
-                  ...router.query,
-                  completeProfileModal: true,
-                },
-              } as any,
-            }),
-            Platform.select({
-              native: "/profile/complete",
-              web: router.asPath,
-            })
-          )
-        }
+        onPress={() => router.push("/profile/edit")}
       >
         {cta}
       </Button>
