+++ packages/app/components/header-dropdown.tsx
@@ -134,22 +134,7 @@ function HeaderDropdown({
 
         <DropdownMenuItem
           onSelect={() => {
-            router.push(
-              Platform.select({
-                native: "/profile/edit",
-                web: {
-                  pathname: router.pathname,
-                  query: {
-                    ...router.query,
-                    editProfileModal: true,
-                  },
-                } as any,
-              }),
-              Platform.select({
-                native: "/profile/edit",
-                web: router.asPath,
-              })
-            );
+            router.push("/profile/edit");
           }}
           key="edit-profile"
         >
