+++ packages/app/components/login/use-login.ts
@@ -19,7 +19,7 @@ export const useLogin = (onLogin?: () => void) => {
   const { rudder } = useRudder();
 
   //#region hooks
-  const { authenticationStatus, logout } = useAuth();
+  const { authenticationStatus, logout, setAuthenticationStatus } = useAuth();
   const {
     loginWithWallet,
     name: walletName,
@@ -35,23 +35,24 @@ export const useLogin = (onLogin?: () => void) => {
   //#endregion
 
   //#region methods
-  const handleLoginFailure = useCallback(function handleLoginFailure(
-    error: any
-  ) {
-    loginSource.current = "undetermined";
+  const handleLoginFailure = useCallback(
+    function handleLoginFailure(error: any) {
+      loginSource.current = "undetermined";
+      setAuthenticationStatus("UNAUTHENTICATED");
+      if (process.env.NODE_ENV === "development" || __DEV__) {
+        console.error(error);
+      }
 
-    if (process.env.NODE_ENV === "development" || __DEV__) {
-      console.error(error);
-    }
-
-    captureException(error, {
-      tags: {
-        login_signature_flow: "use-login.ts",
-        login_magic_link: "use-login.ts",
-      },
-    });
-  },
-  []);
+      captureException(error, {
+        tags: {
+          login_signature_flow: "use-login.ts",
+          login_magic_link: "use-login.ts",
+        },
+      });
+    },
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    []
+  );
 
   const handleSubmitWallet = useCallback(
     async function handleSubmitWallet(params?: SubmitWalletParams) {
@@ -63,8 +64,6 @@ export const useLogin = (onLogin?: () => void) => {
         });
 
         if (isWeb) {
-          console.log(loginSource.current);
-
           params?.onOpenConnectModal?.();
         } else {
           await loginWithWallet();
@@ -82,7 +81,6 @@ export const useLogin = (onLogin?: () => void) => {
         rudder?.track("Button Clicked", {
           name: "Login with email",
         });
-
         return await loginWithEmail(email);
       } catch (error) {
         handleLoginFailure(error);
@@ -123,14 +121,14 @@ export const useLogin = (onLogin?: () => void) => {
   //#region effects
   useStableBlurEffect(handleBlur);
   useEffect(() => {
+    const isAuthenticated = authenticationStatus === "AUTHENTICATED";
     const isLoggedInByMagic =
-      loginSource.current === "magic" &&
-      authenticationStatus === "AUTHENTICATED";
+      loginSource.current === "magic" && isAuthenticated;
     const isLoggedInByWallet =
       loginSource.current === "wallet" && walletStatus === "EXPIRED_NONCE";
 
-    if ((isLoggedInByWallet || isLoggedInByMagic) && onLogin) {
-      onLogin();
+    if (isLoggedInByWallet || isLoggedInByMagic) {
+      onLogin?.();
     }
   }, [authenticationStatus, walletStatus, onLogin]);
